2.3.23: enable DOOM music + fix OPL-induced regressions (LVGL hang, choppy video)
- Music was never playing: prboom d_main.c had a hardcoded nomusicparm=true
  ("Hardcode music and sound disabled -- JD") that short-circuits every music
  call regardless of mus_card. Re-enabled (we now have the OPL backend).
- LVGL hang on DOOM exit: DoomPlayer::Stop hard-kills the engine but never ran
  DOOM's shutdown, so the doom_snd audio task (now on core 0 doing OPL render)
  LEAKED and starved the LCD flush-done crosscore signal -> taskLVGL wedged in
  wait_for_flushing. Stop() now calls I_ShutdownSound() to stop that task (and
  tear down OPL) before lvgl_port_resume. Clean per-launch OPL lifecycle.
- Choppy video (internal SRAM down to 131 B): DBOPL's ~9 KB static wavetables
  sat in internal .bss, eating the headroom the 2.3.18 chunk fix freed. Moved
  to PSRAM via EXT_RAM_BSS_ATTR (read-mostly, OPL runs on idle core 0).
- Fixed %lld stall log (newlib-nano has no long long) -> %lu.

+++ main/video_player.cc
@@ -2254,8 +2254,10 @@ void VideoPlayer::VideoRenderTask(void* arg) {
                     self->dropped_frames_++;
                 }
                 backlog_drops += static_cast<int>(popped);
-                ESP_LOGW(TAG, "Render stalled %lld ms — resync, dropped %lu frames",
-                         (long long)((stall_now_us - last_present_us) / 1000), (unsigned long)popped);
+                // newlib-nano printf has no %lld (long long); use %lu — a stall
+                // duration in ms fits a 32-bit long fine.
+                ESP_LOGW(TAG, "Render stalled %lu ms — resync, dropped %lu frames",
+                         (unsigned long)((stall_now_us - last_present_us) / 1000), (unsigned long)popped);
                 // Anchor audio_clock to the surviving frame so the selector
                 // picks it (it's "future" from the previous audio_clock but we
                 // want it rendered NOW).
